+++ CHANGELOG.md
@@ -21,6 +21,11 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+### New
+
+* Added `Entity` interface to v2 API
+* Added `EntityObject` to v2 API
+
 ## v1.0.1
 
 Released Sunday, 2nd February 2020.